Day 25: Moving down to i64 sped it up even more! Now it is 0.18 seconds.

+++ src/day25.rs
@@ -1,14 +1,14 @@
 use std::mem;
 use num::Integer;
 
-const PUBLIC_SUBJECT: i128 = 7;
-const LOOP_MOD: i128 = 20_201_227;
+const PUBLIC_SUBJECT: i64 = 7;
+const LOOP_MOD: i64 = 20_201_227;
 
 pub fn solve_1() -> String {
     find_encryption_key(19241437, 17346587).to_string()
 }
 
-fn find_encryption_key(card_public: i128, door_public: i128) -> i128 {
+fn find_encryption_key(card_public: i64, door_public: i64) -> i64 {
     let card_loop = find_loop_size(card_public);
     let door_loop = find_loop_size(door_public);
     let guess1 = guess_encryption_key(card_loop, &door_public);
@@ -17,7 +17,7 @@ fn find_encryption_key(card_public: i128, door_public: i128) -> i128 {
     guess1
 }
 
-fn find_loop_size(public_key: i128) -> usize {
+fn find_loop_size(public_key: i64) -> usize {
     Transform::from(PUBLIC_SUBJECT)
         .enumerate()
         .find(|(_, n)| n == &public_key)
@@ -25,24 +25,24 @@ fn find_loop_size(public_key: i128) -> usize {
         .unwrap()
 }
 
-fn guess_encryption_key(device_1_loop: usize, device_2_public_key: &i128) -> i128 {
+fn guess_encryption_key(device_1_loop: usize, device_2_public_key: &i64) -> i64 {
     Transform::from(device_2_public_key.clone()).nth(device_1_loop).unwrap()
 }
 
 #[derive(Debug,Clone)]
 struct Transform {
-    subject_number: i128,
-    value: i128
+    subject_number: i64,
+    value: i64
 }
 
 impl Transform {
-    fn from(subject_number: i128) -> Self {
+    fn from(subject_number: i64) -> Self {
         Transform {subject_number, value: 1}
     }
 }
 
 impl Iterator for Transform {
-    type Item = i128;
+    type Item = i64;
 
     fn next(&mut self) -> Option<Self::Item> {
         let mut future = (&self.value * &self.subject_number).mod_floor(&LOOP_MOD);
